factor out collectNodes
