+++ src/lint/lint.ts
@@ -22,6 +22,7 @@ import {
   getProductions,
   rhsMatches,
 } from './utils';
+import { collectNodes } from './collect-nodes';
 import lintAlgorithmLineEndings from './rules/algorithm-line-endings';
 
 function composeObservers(...observers: Observer[]): Observer {
@@ -62,109 +63,7 @@ export function lint(
   // *******************
   // Walk the whole tree collecting interesting parts
 
-  let grammarNodes: Element[] = [];
-  let grammarParts: string[] = [];
-  let grammarStartOffsets: number[] = [];
-  let grammarLineOffsets: number[] = [];
-
-  let sdos: { grammar: Element; alg: Element }[] = [];
-  let earlyErrors: { grammar: Element; lists: HTMLUListElement[] }[] = [];
-  let algorithms: { element: Element; tree?: EcmarkdownNode }[] = [];
-  let inAnnexB = false;
-  let lintWalker = document.createTreeWalker(document.body, 1 /* elements */);
-  function visitCurrentNode() {
-    let node: Element = lintWalker.currentNode as Element;
-
-    let thisNodeIsAnnexB =
-      node.nodeName === 'EMU-ANNEX' &&
-      node.id === 'sec-additional-ecmascript-features-for-web-browsers';
-    if (thisNodeIsAnnexB) {
-      inAnnexB = true;
-    }
-
-    // Don't bother collecting early errors and SDOs from Annex B.
-    // This is mostly so we don't have to deal with having two inconsistent copies of some of the grammar productions.
-    if (!inAnnexB) {
-      if (node.nodeName === 'EMU-CLAUSE') {
-        // Look for early errors
-        let first = node.firstElementChild;
-        if (first !== null && first.nodeName === 'H1') {
-          let title = first.textContent ?? '';
-          if (title.trim() === 'Static Semantics: Early Errors') {
-            let grammar = null;
-            let lists: HTMLUListElement[] = [];
-            for (let child of (node.children as any) as Iterable<Element>) {
-              if (child.nodeName === 'EMU-GRAMMAR') {
-                if (grammar !== null) {
-                  if (lists.length === 0) {
-                    throw new Error(
-                      'unrecognized structure for early errors: grammar without errors'
-                    );
-                  }
-                  earlyErrors.push({ grammar, lists });
-                }
-                grammar = child;
-                lists = [];
-              } else if (child.nodeName === 'UL') {
-                if (grammar === null) {
-                  throw new Error(
-                    'unrecognized structure for early errors: errors without correspondinig grammar'
-                  );
-                }
-                lists.push(child as HTMLUListElement);
-              }
-            }
-            if (grammar === null) {
-              throw new Error('unrecognized structure for early errors: no grammars');
-            }
-            if (lists.length === 0) {
-              throw new Error('unrecognized structure for early errors: grammar without errors');
-            }
-            earlyErrors.push({ grammar, lists });
-          }
-        }
-      } else if (node.nodeName === 'EMU-GRAMMAR') {
-        // Look for grammar definitions and SDOs
-        if (node.getAttribute('type') === 'definition') {
-          let loc = getLocation(dom, node);
-          let start = loc.startTag.endOffset;
-          let end = loc.endTag.startOffset;
-          grammarStartOffsets.push(start);
-          grammarLineOffsets.push(loc.startTag.line);
-          let realSource = sourceText.slice(start, end);
-          grammarParts.push(realSource);
-          grammarNodes.push(node as Element);
-        } else if (node.getAttribute('type') !== 'example') {
-          let next = lintWalker.nextSibling() as Element;
-          if (next) {
-            if (next.nodeName === 'EMU-ALG') {
-              sdos.push({ grammar: node, alg: next });
-            }
-            lintWalker.previousSibling();
-          }
-        }
-      }
-    }
-
-    if (node.nodeName === 'EMU-ALG' && node.getAttribute('type') !== 'example') {
-      algorithms.push({ element: node });
-    }
-
-    let firstChild = lintWalker.firstChild();
-    if (firstChild) {
-      while (true) {
-        visitCurrentNode();
-        let next = lintWalker.nextSibling();
-        if (!next) break;
-      }
-      lintWalker.parentNode();
-    }
-
-    if (thisNodeIsAnnexB) {
-      inAnnexB = false;
-    }
-  }
-  visitCurrentNode();
+  let { mainGrammar, sdos, earlyErrors, algorithms } = collectNodes(sourceText, dom, document);
 
   // *******************
   // Parse the grammar with Grammarkdown and collect its diagnostics, if any
@@ -174,10 +73,10 @@ export function lint(
   let fakeHost: SyncHost = {
     readFileSync(file: string) {
       let idx = parseInt(file);
-      if (idx.toString() !== file || idx < 0 || idx >= grammarParts.length) {
+      if (idx.toString() !== file || idx < 0 || idx >= mainGrammar.length) {
         throw new Error('tried to read non-existent ' + file);
       }
-      return grammarParts[idx];
+      return mainGrammar[idx].source;
     },
     resolveFile(file: string) {
       return file;
@@ -194,7 +93,7 @@ export function lint(
     noChecks: false,
     noUnusedParameters: true,
   };
-  let grammar = new GrammarFile(Object.keys(grammarParts), compilerOptions, fakeHost);
+  let grammar = new GrammarFile(Object.keys(mainGrammar), compilerOptions, fakeHost);
   grammar.parseSync();
   grammar.checkSync();
 
@@ -207,7 +106,7 @@ export function lint(
       .getDiagnosticInfos({ formatMessage: true, detailedMessage: false })
       .forEach(m => {
         let idx = +m.sourceFile!.filename;
-        let grammarLoc = getLocation(dom, grammarNodes[idx]);
+        let grammarLoc = getLocation(dom, mainGrammar[idx].element);
 
         let { line, column } = grammarkdownLocationToTrueLocation(
           grammarLoc,
@@ -385,7 +284,7 @@ export function lint(
 
   grammar.emitSync(undefined, (file, source) => {
     let name = +file.split('.')[0];
-    let node = grammarNodes[name];
+    let node = mainGrammar[name].element;
     if ('grammarkdownOut' in node) {
       throw new Error('unexpectedly regenerating grammarkdown output for node ' + name);
     }
